+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 - Dependencies - Updated Meziantou.Analyzer to 2.0.179
 - SDK - Updated DotNet SDK to 9.0.100
 - Dependencies - Updated FunFair.Test.Common to 6.1.239.979
+- Dependencies - Updated Microsoft.Extensions to 9.0.0
 ### Removed
 ### Deployment Changes